relocate misplaced IT test
